+++ semana12/labeddit/src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AuthProvider from './context/AuthProvider';
+import AuthProvider from './contexts/AuthProvider';
 import Routes from './Routes';
 
 const App = () => (
